+++ src/main.js
@@ -2307,7 +2307,9 @@ async function joinAll(
                 const __fileId = _filesWithSlices.pop();
                 const __file = _files[__fileId];
                 const exploded = splitByOtSlices({}, __file.meta.id, false, 'ot', [], false, true);
-                _files.splice(__fileId, 1, ...exploded);
+                if (Array.isArray(exploded)) {
+                    _files.splice(__fileId, 1, ...exploded);
+                }
             }
         }
 
@@ -2601,6 +2603,7 @@ async function joinAll(
         }
     } catch (joinError) {
         let errorMessage = `An unexpected error was encountered while building the sample chain(s).`;
+        debugger;
         if (joinError) {
             errorMessage += `\n\n "${joinError.toString ? joinError.toString() : JSON.stringify(joinError)}"`;
         }
@@ -3905,7 +3908,7 @@ const buildRowMarkupFromFile = (f, type = 'main') => {
       `</td>
       <td class="split-td">
           <button title="Slice sample (Ctrl/Cmd + Click to Clear Slice Data)" onpointerdown="digichain.splitAction(event, '${f.meta.id}')" class="button-clear split gg-menu-grid-r ${metaFiles.getByFile(
-        f)?.cssClass}" data-slice-count="${f.meta?.slices?.length || f.meta?.op1Json?.sliceCount || ''}" data-has-invalid-slice="${Array.isArray(f.meta?.slices)? (f.meta.slices.some(s => s.e - s.s < 5) ? '1' : '') : ''}"><i class="gg-menu-grid-r has-ctrl-mod-i"></i></button>
+        f)?.cssClass}" data-slice-count="${f.meta?.slices?.length || f.meta?.op1Json?.sliceCount || ''}" data-has-invalid-slice="${f.meta?.slices?.length ? f.meta.slices.findIndex(s => s.e - s.s < 2) + 1 : '0'}"><i class="gg-menu-grid-r has-ctrl-mod-i"></i></button>
       </td>
       <td class="duplicate-td">
           <button title="Duplicate sample." onpointerdown="digichain.duplicate(event, '${f.meta.id}')" class="button-clear duplicate"><i class="gg-duplicate has-shift-mod-i"></i></button>