+++ changelog.md
@@ -3,7 +3,8 @@
 - Updated note detection to find notes from Logic Pro generated filenames.
 - Sorting by note on Logic Pro generated AIFF files now correctly sorts.
 - Fix session export files have .dcsd.dcsd as the filename extension in the saved notification text.
-- Add !!! text as the slice count when there is an invalid slice on the file.
+- Add !!-s text as the slice count when there is an invalid slice on the file where s is the first encountered invalid slice number.
+- Graceful handling of invalid slices when joining, if an invalid slice is encountered, the join will use the full source sample instead.
 
 2025.12.07 1.5.2
 - Corrected inconsistent import ordering in macOS file drops.